fix: refactor assert functions
